Move XCUIElement utils to dedicated file in UITestsFoundation
