+++ WordPress/UITestsFoundation/XCUIElement+Utils.swift
@@ -1,0 +1,53 @@
+import XCTest
+
+// TODO: This should go XCUITestHelpers if not there already
+public extension XCUIElement {
+
+    /**
+     Pastes text from clipboard to the field
+     Useful for scenarios where typing is problematic, e.g. secure text fields in Russian.
+     - Parameter text: the text to paste into the field
+     */
+    func pasteText(_ text: String) -> Void {
+        let previousPasteboardContents = UIPasteboard.general.string
+        UIPasteboard.general.string = text
+
+        press(forDuration: 1.2)
+        XCUIApplication().menuItems.firstMatch.tap()
+
+        if let string = previousPasteboardContents {
+            UIPasteboard.general.string = string
+        }
+    }
+
+    @discardableResult
+    // TODO: When moving to framework, find name that doesn't trigger grammar warning
+    func waitForHittability(timeout: TimeInterval) -> Bool {
+
+        let predicate = NSPredicate(format: "isHittable == true")
+        let elementPredicate = XCTNSPredicateExpectation(predicate: predicate, object: self)
+        let result = XCTWaiter.wait(for: [elementPredicate], timeout: timeout)
+
+        return result == .completed
+    }
+
+    var isFullyVisibleOnScreen: Bool {
+        guard self.exists && !self.frame.isEmpty && self.isHittable else { return false }
+        return XCUIApplication().windows.element(boundBy: 0).frame.contains(self.frame)
+    }
+
+    func scroll(byDeltaX deltaX: CGFloat, deltaY: CGFloat) {
+        let startCoordinate = self.coordinate(withNormalizedOffset: CGVector(dx: 0, dy: 0))
+        let destination = startCoordinate.withOffset(CGVector(dx: deltaX, dy: deltaY * -1))
+
+        startCoordinate.press(forDuration: 0.01, thenDragTo: destination)
+    }
+
+    /// Removes any current text in the field
+    func clearTextIfNeeded() -> Void {
+        let app = XCUIApplication()
+
+        self.press(forDuration: 1.2)
+        app.keys["delete"].tap()
+    }
+}
